requirement.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-bootstrap-form==3.4
django-crispy-forms==1.11.1
django-field-history==0.8.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-widget-tweaks==1.4.8
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
swapper==1.1.2.post1
twilio==6.58.0
urllib3==1.26.4
